extract the paypal models out of the controller to remove noise
