+++ app/controllers/PaypalController.scala
@@ -6,24 +6,19 @@ import cats.syntax.show._
 import cookies.ContribTimestampCookieAttributes
 import cookies.syntax._
 import com.gu.i18n.{CountryGroup, Currency}
-import com.netaporter.uri.Uri
 import com.paypal.api.payments.Payment
+import controllers.forms.{AuthRequest, AuthResponse}
 import models._
 import monitoring._
-import play.api.libs.ws.WSClient
 import play.api.mvc._
 import services.PaymentServices
 import play.api.data.Form
 import utils.MaxAmount
 import play.api.libs.json._
-import play.api.libs.json.Reads._
-import play.api.libs.functional.syntax._
 import play.api.data.Forms._
-import play.api.http.Writeable
 import play.filters.csrf.CSRFCheck
 
 import scala.concurrent.{ExecutionContext, Future}
-import scala.util.Try
 
 class PaypalController(paymentServices: PaymentServices, checkToken: CSRFCheck, cloudWatchMetrics: CloudWatchMetrics)(implicit ec: ExecutionContext)
   extends Controller with Redirect with TagAwareLogger with LoggingTagsProvider {
@@ -90,83 +85,6 @@ class PaypalController(paymentServices: PaymentServices, checkToken: CSRFCheck, 
     paypalService.executePayment(paymentId, payerId)
       .map { payment => storeMetaData(payment); payment }
       .fold(notOkResult, okResult)
-  }
-
-  case class AuthRequest private (
-    countryGroup: CountryGroup,
-    amount: BigDecimal,
-    cmp: Option[String],
-    intCmp: Option[String],
-    refererPageviewId: Option[String],
-    refererUrl: Option[String],
-    ophanPageviewId: Option[String],
-    ophanBrowserId: Option[String],
-    ophanVisitId: Option[String]
-  )
-
-  object AuthRequest {
-    /**
-      * We need to ensure there's no fragment in the URL here, as PayPal appends some query parameters to the end of it,
-      * which will be removed by the browser (due to the URL stripping rules) in its requests.
-      *
-      * See: https://www.w3.org/TR/referrer-policy/#strip-url
-      *
-      */
-    def withSafeRefererUrl(
-                            countryGroup: CountryGroup,
-                            amount: BigDecimal,
-                            cmp: Option[String],
-                            intCmp: Option[String],
-                            refererPageviewId: Option[String],
-                            refererUrl: Option[String],
-                            ophanPageviewId: Option[String],
-                            ophanBrowserId: Option[String],
-                            ophanVisitId: Option[String]
-                          ): AuthRequest = {
-      val safeRefererUrl = refererUrl.flatMap(url => Try(Uri.parse(url).copy(fragment = None).toString).toOption)
-
-      new AuthRequest(countryGroup, amount, cmp, intCmp, refererPageviewId, safeRefererUrl, ophanPageviewId, ophanBrowserId, ophanVisitId)
-    }
-
-
-    implicit val authRequestReads: Reads[AuthRequest] = (
-      (__ \ "countryGroup").read[CountryGroup] and
-        (__ \ "amount").read(min[BigDecimal](1)) and
-        (__ \ "cmp").readNullable[String] and
-        (__ \ "intCmp").readNullable[String] and
-        (__ \ "refererPageviewId").readNullable[String] and
-        (__ \ "refererUrl").readNullable[String] and
-        (__ \ "ophanPageviewId").readNullable[String] and
-        (__ \ "ophanBrowserId").readNullable[String] and
-        (__ \ "ophanVisitId").readNullable[String]
-      ) (AuthRequest.withSafeRefererUrl _)
-  }
-
-  case class AuthResponse(approvalUrl: Uri, paymentId: String)
-
-  object AuthResponse {
-    import cats.syntax.either._
-    import scala.collection.JavaConverters._
-
-    def fromPayment(payment: Payment): Either[String, AuthResponse] = Either.fromOption(for {
-        links <- Option(payment.getLinks)
-        approvalLinks <- links.asScala.find(_.getRel.equalsIgnoreCase("approval_url"))
-        approvalUrl <- Option(approvalLinks.getHref)
-        paymentId <- Option(payment.getId)
-      } yield AuthResponse(Uri.parse(approvalUrl), paymentId), "Unable to parse payment")
-  }
-
-  implicit val UriWrites = new Writes[Uri] {
-    override def writes(uri: Uri): JsValue = JsString(uri.toString)
-  }
-
-  implicit val AuthResponseWrites = Json.writes[AuthResponse]
-
-  implicit val CountryGroupReads = new Reads[CountryGroup] {
-    override def reads(json: JsValue): JsResult[CountryGroup] = json match {
-      case JsString(id) => CountryGroup.byId(id).map(JsSuccess(_)).getOrElse(JsError("invalid CountryGroup id"))
-      case _ => JsError("invalid value for country group")
-    }
   }
 
   private def capAmount(amount: BigDecimal, currency: Currency): BigDecimal = amount min MaxAmount.forCurrency(currency)
