+++ app/utils/JsonUtils.scala
@@ -1,0 +1,13 @@
+package utils
+
+import com.gu.i18n.CountryGroup
+import play.api.libs.json._
+
+object JsonUtils {
+  implicit val countryGroupReads = new Reads[CountryGroup] {
+    override def reads(json: JsValue): JsResult[CountryGroup] = json match {
+      case JsString(id) => CountryGroup.byId(id).map(JsSuccess(_)).getOrElse(JsError("invalid CountryGroup id"))
+      case _ => JsError("invalid value for country group")
+    }
+  }
+}
