+++ conf/application.conf
@@ -1,5 +1,5 @@
 play.application.loader = "wiring.AppLoader"
-
+play.http.secret.key="a-very-secret-dev-key"
 play.http.session.secure=true
 
 play.http.errorHandler = "monitoring.ErrorHandler"
